Rename combat skills module
