+++ combat/scripts/skills.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Optional, Union
 
 from ..combat_actions import SkillAction
-from ..combat_skills import SKILL_CLASSES, Skill
+from ..skills import SKILL_CLASSES, Skill
 from world.skills.utils import maybe_start_combat
